remove unused Foundation import
